server/requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
pandas
python-dotenv
google-generativeai
